Move ExecEnvironment and Formatter to lib/runtime
Reflecting their use
