+++ index.js
@@ -12,12 +12,12 @@
 const Ast = require('./lib/ast');
 const Compiler = require('./lib/compiler');
 const Grammar = require('./lib/grammar_api');
-const ExecEnvironment = require('./lib/exec_environment');
+const ExecEnvironment = require('./lib/runtime/exec_environment');
 const Type = require('./lib/type');
 const SchemaRetriever = require('./lib/schema');
 const Generate = require('./lib/generate');
 const Describe = require('./lib/describe');
-const Formatter = require('./lib/formatter');
+const Formatter = require('./lib/runtime/formatter');
 const PermissionChecker = require('./lib/permission_checker');
 const Units = require('./lib/units');
 const NNSyntax = require('./lib/nn_syntax');
