+++ lib/runtime/formatter.js
@@ -11,8 +11,8 @@
 
 const vm = require('vm');
 
-const { clean } = require('./utils');
-const FormatUtils = require('./format_utils');
+const { clean } = require('../utils');
+const FormatUtils = require('../format_utils');
 
 function compileCode(code) {
     return vm.runInNewContext(code);
